refactor tasks page into modular components
